Removed auto& in GameObject, fixed an exceptions bug in main

+++ src/Player.cpp
@@ -34,7 +34,7 @@ namespace
 	{
 		int dirX;
 
-		AircraftMoverX(float x) :
+		AircraftMoverX(int x) :
 			dirX(x)
 		{
 		}
